% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{build_query}
\alias{build_query}
\title{Given a named character vector, craft a query}
\usage{
build_query(x, elements = NULL)
}
\arguments{
\item{x}{a named character vector}

\item{elements}{a character vector of elements to include}
}
\value{
character query string
}
\description{
Given a named character vector, craft a query
}
\examples{
\dontrun{
   s <- c(lat = '45.0', lon = '-77', Unit = 'm') 
   elements <- c("mint", "maxt")
   build_query(s, elements = elements)
   [1] "lat=45.0&lon=-77&Unit=m&mint=mint&maxt=maxt"
}
}

