% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{build_query_element}
\alias{build_query_element}
\title{Given an element list of names from standard list found}
\usage{
build_query_element(x = c("mint", "max", "temp"))
}
\arguments{
\item{x}{a character vector of element names}
}
\value{
query character string
}
\description{
Given an element list of names from standard list found
}
\seealso{
\url{http://graphical.weather.gov/xml/docs/elementInputNames.php}
}

