% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{check_response}
\alias{check_response}
\title{Check an \code{httr::response} class object and possible pare to \code{XML::xmlNode}}
\usage{
check_response(rsp, encoding = "UTF-8")
}
\arguments{
\item{rsp}{\code{httr::response} class object}

\item{encoding}{character, by defalt 'UTF-8'}
}
\value{
\code{XML::xmlNode} object
}
\description{
Check an \code{httr::response} class object and possible pare to \code{XML::xmlNode}
}

