% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWML_get_time_layout}
\alias{DWML_get_time_layout}
\title{Retrieve a list of time layout data frames}
\usage{
DWML_get_time_layout(obj, key = NULL, form = c("POSIXct", "character")[1])
}
\arguments{
\item{obj}{the DWMLDataRefClass object}

\item{key}{charcater vector of one or more time-layout-keys or NULL to get all}

\item{form}{character the time format to retrieve}
}
\value{
a list of one or more data frames
}
\description{
Retrieve a list of time layout data frames
}

