% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Base.R
\docType{class}
\name{DWMLBaseRefClass-class}
\alias{DWMLBaseRefClass}
\alias{DWMLBaseRefClass-class}
\title{The root class for any DWML XML structure}
\description{
The root class for any DWML XML structure
}
\section{Fields}{

\describe{
\item{\code{node}}{a single XML::xmlNode}
}}
\seealso{
\url{http://graphical.weather.gov/xml/mdl/XML/Design/MDL_XML_Design.pdf}
}

