% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{class}
\name{DWMLDataRefClass-class}
\alias{DWMLDataRefClass}
\alias{DWMLDataRefClass-class}
\title{A DWML class for for the data section}
\description{
A DWML class for for the data section
}
\section{Fields}{

\describe{
\item{\code{location}}{a data frame of location information}
}}

