% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_common_starts}
\alias{DWMLDataRefClass_common_starts}
\title{Retrieves the common start times across all time-layouts}
\value{
a named vector of POSIXct or possibly empty
}
\description{
Retrieves the common start times across all time-layouts
}

