% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_get_by_time}
\alias{DWMLDataRefClass_get_by_time}
\title{Get data by time-layout}
\arguments{
\item{key}{character string - pattern matched so if you have series named
k-p24h-n7-1, k-p24h-n7-2 and k-p24h-n7-3 then requesting
k-p24h-n7-3 will get the third, but k-p24h-n7 will get all in the series}
}
\value{
data frame
}
\description{
Get data by time-layout
}

