% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_get_data}
\alias{DWMLDataRefClass_get_data}
\title{Retrieve the data by name}
\arguments{
\item{name}{character of the parameter to retrieve}

\item{by}{character data can be assmbled by time or by location}
}
\value{
data.frame with zero or more rows
}
\description{
Retrieve the data by name
}

