% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_get_time_layout}
\alias{DWMLDataRefClass_get_time_layout}
\title{Extracts time-layout information}
\value{
a list of data.frames - one per layout-key
}
\description{
The <time-layout> element {dw:time-layoutType } [+] contains the start and 
 stop valid times and any associated period names for the data. Since 
 different environmental parameters have different time schemes (valid at 
 different interval and available for different lengths of time into the 
 future), there will be one <time-layout> element for each of these unique 
 temporal configurations. Each data parameter will reference exactly one of 
 these time layouts (R2.2.3).
}
\details{
time-layout
<time-layout time-coordinate="local" summarization="none">
  <layout-key>k-p24h-n6-2</layout-key>
  <start-valid-time>2016-02-26T19:00:00-05:00</start-valid-time>
  <end-valid-time>2016-02-27T08:00:00-05:00</end-valid-time>
  <start-valid-time>2016-02-27T19:00:00-05:00</start-valid-time>
  <end-valid-time>2016-02-28T08:00:00-05:00</end-valid-time>
  ...
  <start-valid-time>2016-03-02T19:00:00-05:00</start-valid-time>
  <end-valid-time>2016-03-03T08:00:00-05:00</end-valid-time>
</time-layout>
}

