% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_list_time_layout}
\alias{DWMLDataRefClass_list_time_layout}
\title{Retrieves a character vector of time-layout keys}
\value{
character vector - possibly empty

list of character vectors - possibly empty
}
\description{
Retrieves a character vector of time-layout keys

Retrieves a list of character vectors describing the parameters
}

