% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{DWMLDataRefClass_time_layout}
\alias{DWMLDataRefClass_time_layout}
\title{Extracts paramaters information}
\value{
a list of data.frames - one per appplicable-location
}
\description{
The <time-layout> element {dw:time-layoutType } [+] contains the start and 
 stop valid times and any associated period names for the data. Since 
 different environmental parameters have different time schemes (valid at 
 different interval and available for different lengths of time into the 
 future), there will be one <time-layout> element for each of these unique 
 temporal configurations. Each data parameter will reference exactly one of 
 these time layouts (R2.2.3).
<parameters applicable-location="point1">
 <temperature type="maximum" units="Fahrenheit" time-layout="k-p24h-n7-1">
  <name>Daily Maximum Temperature</name>
  <value>51</value>
   ...
  <value>53</value>
 </temperature>
  <temperature type="minimum" units="Fahrenheit" time-layout="k-p24h-n6-2">
  <name>Daily Minimum Temperature</name>
  <value>30</value>
  <value>34</value>
  ...
  <value>38</value>
 </temperature>
</parameters>
}

