% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Head.R
\docType{class}
\name{DWMLHeadRefClass-class}
\alias{DWMLHeadRefClass}
\alias{DWMLHeadRefClass-class}
\title{A DWML class for for the head section}
\description{
The head section doesn't provide much dynamic info so we have only a 
very simple behavior associated with it.
}

