% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Head.R
\name{DWMLHeadRefClass_concise_name}
\alias{DWMLHeadRefClass_concise_name}
\title{Retrieve the concise name of the data set}
\value{
character value of the concise-name attribute
}
\description{
Retrieve the concise name of the data set
}

