% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Head.R
\name{DWMLHeadRefClass_creation_date}
\alias{DWMLHeadRefClass_creation_date}
\title{Retrieve the creation date of the data set ala '2016-02-26T13:25:37Z'}
\arguments{
\item{form}{character either 'character' or 'POSIXct'}
}
\value{
character of POSIXct value of the creation date
}
\description{
Retrieve the creation date of the data set ala '2016-02-26T13:25:37Z'
}

