% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLonList.R
\docType{class}
\name{DWMLLatLonListRefClass-class}
\alias{DWMLLatLonListRefClass}
\alias{DWMLLatLonListRefClass-class}
\title{A DWML class for for the lonLatList section}
\description{
A DWML class for for the lonLatList section
}

