% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatLonList.R
\name{DWMLLatLonListRefClass_extract_location}
\alias{DWMLLatLonListRefClass_extract_location}
\title{Extracts location information}
\arguments{
\item{form}{character specifiction for as_is, character or numeric output}
}
\value{
character or data frame, possibly with zero rows
}
\description{
The \code{as_is} value for the \code{form} argument is useful for creating
 a data query for multiple points
}

