% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top.R
\docType{class}
\name{DWMLTopRefClass-class}
\alias{DWMLTopRefClass}
\alias{DWMLTopRefClass-class}
\title{The top level DWML class for containing a non-exception query response that contains
 a head and a data section.}
\description{
The top level DWML class for containing a non-exception query response that contains
 a head and a data section.
}
\section{Fields}{

\describe{
\item{\code{version}}{character version string}

\item{\code{head}}{DWMLHeadRefClass container}

\item{\code{data}}{DWMLDataRefClass contaier}
}}
\seealso{
\url{http://graphical.weather.gov/xml/mdl/XML/Design/MDL_XML_Design.pdf}
}

