% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top.R
\name{DWMLTopRefClass_get_data}
\alias{DWMLTopRefClass_get_data}
\title{Get data by name and time-layout key}
\arguments{
\item{...}{arguments for DWMLDataRefClass$get_data()}
}
\value{
data.frame or NULL
}
\description{
Get data by name and time-layout key
}

