% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Top.R
\name{DWMLTopRefClass_get_location}
\alias{DWMLTopRefClass_get_location}
\title{Get the location data if available}
\arguments{
\item{...}{further arguments for DWMLDataRefClass$get_location()}
}
\value{
data frame or NULL
}
\description{
Get the location data if available
}

