% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_ndfdXMLclient_basic}
\alias{get_ndfdXMLclient_basic}
\title{This function is used during development to create and save the data object
 \code{ndfdXMLclient_basic}.}
\usage{
get_ndfdXMLclient_basic()
}
\value{
a named chacter vector of basic data elements for queries
}
\description{
This function is used during development to create and save the data object
 \code{ndfdXMLclient_basic}.
}

