% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_ndfdXMLclient_groups}
\alias{get_ndfdXMLclient_groups}
\title{This function is used during development to create and save the data object
 ndfdXMLclient_groups.}
\usage{
get_ndfdXMLclient_groups()
}
\value{
a named list of query elements for various query types
}
\description{
This function is used during development to create and save the data object
 ndfdXMLclient_groups.
}

