% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_ndfdXMLclient_vars}
\alias{get_ndfdXMLclient_vars}
\title{This function is used during development to create and save the data object
 \code{ndfdXMLclient_vars}.}
\usage{
get_ndfdXMLclient_vars()
}
\value{
a named list of default variable values
}
\description{
This function is used during development to create and save the data object
 \code{ndfdXMLclient_vars}.
}

