% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_query}
\alias{get_query}
\title{Retrieve a response for a given query}
\usage{
get_query(query,
  baseuri = "http://graphical.weather.gov/xml/sample_products/browser_interface",
  interface = c("ndfdXMLclient.php", "ndfdBrowserClientByDay")[1])
}
\arguments{
\item{query}{character string}

\item{baseuri}{character the base URI}

\item{interface}{the interface to use with the baseuri}
}
\value{
object inheriting DWMLBaseRefClass
}
\description{
Retrieve a response for a given query
}

