% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_exception}
\alias{is_exception}
\title{Test xml_node or NodeRefClass is an exception}
\usage{
is_exception(x, space = "exc")
}
\arguments{
\item{x}{node object to test}

\item{space}{the namespace to test}
}
\value{
logical
}
\description{
Test xml_node or NodeRefClass is an exception
}

