% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_xml_node}
\alias{is_xml_node}
\title{Test if an object inherits from xml2::xml_node}
\usage{
is_xml_node(x, classname = "xml_node")
}
\arguments{
\item{x}{object to test}

\item{classname}{character, the class name to test against, by default 'XMLAbstractNode'}
}
\value{
logical
}
\description{
Test if an object inherits from xml2::xml_node
}

