% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_xmlNode}
\alias{is_xmlNode}
\title{Test if an object inherits from XML::XMLAbstractNode}
\usage{
is_xmlNode(x, classname = "XMLAbstractNode")
}
\arguments{
\item{x}{object to test}

\item{classname}{character, the class name to test against, by default 'XMLAbstractNode'}
}
\value{
logical
}
\description{
Test if an object inherits from XML::XMLAbstractNode
}

