% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{list_parameter_keys}
\alias{list_parameter_keys}
\title{Retrieve a list of time layout keys for parameters}
\usage{
list_parameter_keys(x)
}
\arguments{
\item{x}{DWMLDataRefClass or XMLNode}
}
\value{
a named list of time-layout keys for parameters
}
\description{
Retrieve a list of time layout keys for parameters
}

