% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{list_this}
\alias{list_this}
\title{Constructs a query to list resources available}
\usage{
list_this(what = "points_in_subgrid", ...)
}
\arguments{
\item{what}{character - the type of query which defines the defaults.  Defaults
for each type of query are identified \code{ndfdXMLclient_vars} and 
\code{ndfdXMLclient_groups}
\itemize{
    \item points_in_subgrid, A List of NDFD Points for a Subgrid
    \item points_on_line, A List of NDFD Points for a Line
    \item zipcodes, A List of NDFD Points for a Zipcode
    \item cities, A List of NDFD Points for a List of Cities
    \item centerpoint, A List of NDFD Points for a Subgrid Defined by a Center Point
    \item corners, A List of NDFD Points for the Corners of an NDFD Grid
}}
}
\description{
Defaults are stored in \code{ndfdXMLclient_vars} and are accessed by 
 the type of resources requested.  For instance the defaults for a 
 \code{points_in_subgrid} are found in \code{ndfdXMLclient_groups} 
 under \code{list_points_in_subgrid}
}
\examples{
\dontrun{
# default for points, and then with lon and elements specified
list_this(what = "points_in_subgrid")
query_this(what = "point_in_subgrid", lon = -69)
# query by zipcode
query_this(what = "zipcode", zipCodeList = "04096",element = c("dew", "temp", "snow"))
}
}
\seealso{
\url{http://graphical.weather.gov/xml/rest.php#XML_contents}
}

