% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{list_time_layout_keys}
\alias{list_time_layout_keys}
\title{Retrieve a list of the time-layout keys in a DWMLDataRefClass}
\usage{
list_time_layout_keys(x)
}
\arguments{
\item{x}{DWMLDataRefClass or XMLNode}
}
\value{
character vector or NULL
}
\description{
Retrieve a list of the time-layout keys in a DWMLDataRefClass
}

