% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n2c}
\alias{n2c}
\title{Convert numeric values to character}
\usage{
n2c(x, fmt = "\%0.5f")
}
\arguments{
\item{x}{numeric or character value to format}

\item{fmt}{character format}
}
\value{
character representation of input
}
\description{
Convert numeric values to character
}

