% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfd-package.R
\docType{package}
\name{ndfd-package}
\alias{ndfd}
\alias{ndfd-package}
\title{ndfd-package}
\description{
A RESTful R interface to the NWS National Digital Forecast Database#' @seealso \url{http://graphical.weather.gov/xml/rest.php}
}

