% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{ndfd_uri}
\alias{ndfd_uri}
\title{Construct a NDFD uri possibly with a query}
\usage{
ndfd_uri(query = NULL,
  baseuri = "http://graphical.weather.gov/xml/sample_products/browser_interface",
  interface = c("ndfdXMLclient.php", "ndfdBrowserClientByDay")[1])
}
\arguments{
\item{query}{character string}

\item{baseuri}{character the base URI}

\item{interface}{the interface to use with the baseuri}
}
\value{
uri
}
\description{
Construct a NDFD uri possibly with a query
}

