% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfd-package.R
\docType{data}
\name{ndfdElementInputNames}
\alias{ndfdElementInputNames}
\title{The available input elements for queries}
\format{a named 55 element character vector
\describe{
 \item{maxt}{Maximum Temperature }
 \item{mint}{Minimum Temperature }
 \item{temp}{3 Hourly Temperature }
 \item{dew}{Dewpoint Temperature }
 \item{appt}{Apparent Temperature }
 \item{pop12}{12 Hour Probability of Precipitation }
 \item{qpf}{Liquid Precipitation Amount }
 \item{snow}{Snowfall Amount }
 \item{sky}{Cloud Cover Amount }
 \item{rh}{Relative Humidity }
 \item{wspd}{Wind Speed }
 \item{wdir}{Wind Direction }
 \item{wx}{Weather }
 \item{icons}{Weather Icons }
 \item{waveh}{Wave Height }
 \item{incw34}{Probabilistic Tropical Cyclone Wind Speed >34 Knots (Incremental) }
 \item{incw50}{Probabilistic Tropical Cyclone Wind Speed >50 Knots (Incremental) }
 \item{incw64}{Probabilistic Tropical Cyclone Wind Speed >64 Knots (Incremental) }
 \item{cumw34}{Probabilistic Tropical Cyclone Wind Speed >34 Knots (Cumulative) }
 \item{cumw50}{Probabilistic Tropical Cyclone Wind Speed >50 Knots (Cumulative) }
 \item{cumw64}{Probabilistic Tropical Cyclone Wind Speed >64 Knots (Cumulative) }
 \item{wgust}{Wind Gust }
 \item{critfireo}{Fire Weather from Wind and Relative Humidity }
 \item{dryfireo}{Fire Weather from Dry Thunderstorms }
 \item{conhazo}{Convective Hazard Outlook }
 \item{ptornado}{Probability of Tornadoes }
 \item{phail}{Probability of Hail }
 \item{ptstmwinds}{Probability of Damaging Thunderstorm Winds }
 \item{pxtornado}{Probability of Extreme Tornadoes }
 \item{pxhail}{Probability of Extreme Hail }
 \item{pxtstmwinds}{Probability of Extreme Thunderstorm Winds }
 \item{ptotsvrtstm}{Probability of Severe Thunderstorms }
 \item{pxtotsvrtstm}{Probability of Extreme Severe Thunderstorms }
 \item{tmpabv14d}{Probability of 8- To 14-Day Average Temperature Above Normal }
 \item{tmpblw14d}{Probability of 8- To 14-Day Average Temperature Below Normal }
 \item{tmpabv30d}{Probability of One-Month Average Temperature Above Normal }
 \item{tmpblw30d}{Probability of One-Month Average Temperature Below Normal }
 \item{tmpabv90d}{Probability of Three-Month Average Temperature Above Normal }
 \item{tmpblw90d}{Probability of Three-Month Average Temperature Below Normal }
 \item{prcpabv14d}{Probability of 8- To 14-Day Total Precipitation Above Median }
 \item{prcpblw14d}{Probability of 8- To 14-Day Total Precipitation Below Median }
 \item{prcpabv30d}{Probability of One-Month Total Precipitation Above Median }
 \item{prcpblw30d}{Probability of One-Month Total Precipitation Below Median }
 \item{prcpabv90d}{Probability of Three-Month Total Precipitation Above Median }
 \item{prcpblw90d}{Probability of Three-Month Total Precipitation Below Median }
 \item{precipa_r}{Real-time Mesoscale Analysis Precipitation }
 \item{sky_r}{Real-time Mesoscale Analysis GOES Effective Cloud Amount }
 \item{td_r}{Real-time Mesoscale Analysis Dewpoint Temperature }
 \item{temp_r}{Real-time Mesoscale Analysis Temperature }
 \item{wdir_r}{Real-time Mesoscale Analysis Wind Direction }
 \item{wspd_r}{Real-time Mesoscale Analysis Wind Speed }
 \item{wwa}{Watches, Warnings, and Advisories }
 \item{iceaccum}{Ice Accumulation }
 \item{maxrh}{Maximum Relative Humidity }
 \item{minrh}{Minimum Relative Humidity }
}}
\source{
\url{http://graphical.weather.gov/xml/docs/elementInputNames.php}
}
\usage{
ndfdElementInputNames
}
\description{
The available input elements for queries
}
\keyword{datasets}

