% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDFD.R
\docType{class}
\name{NDFDRefClass-class}
\alias{NDFDRefClass}
\alias{NDFDRefClass-class}
\title{A DWML generic use class}
\description{
A DWML generic use class
}
\section{Fields}{

\describe{
\item{\code{uri}}{character}
}}

