% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfd-package.R
\docType{data}
\name{ndfdXMLclient_basic}
\alias{ndfdXMLclient_basic}
\title{The basic query terms for most searches}
\format{a character vector of 4 elements.
\itemize{
 \item product
 \item begin
 \item end
 \item Unit
 }}
\source{
\url{http://graphical.weather.gov/xml/rest.php#XML_contents}
}
\usage{
ndfdXMLclient_basic
}
\description{
The basic query terms for most searches
}
\keyword{datasets}

