% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfd-package.R
\docType{data}
\name{ndfdXMLclient_groups}
\alias{ndfdXMLclient_groups}
\title{Query search terms groups by the type of search}
\format{a named list of character vectors
\itemize{
 \item query_point
 \item query_multipoint
 \item query_subgrid
 \item list_points_in_subgrid
 \item query_line
 \item list_points_on_line
 \item query_zipcodes
 \item list_zipcodes
 \item query_cities
 \item list_cities
 \item query_centerpoint
 \item list_centerpoint
 \item list_corners
 \item query_single_time
 }}
\source{
\url{http://graphical.weather.gov/xml/rest.php#XML_contents}
}
\usage{
ndfdXMLclient_groups
}
\description{
It is best to study the source and the structure of this list to get a sense
 of the types of queries and the related search terms.
}
\keyword{datasets}

