% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndfd-package.R
\docType{data}
\name{ndfdXMLclient_vars}
\alias{ndfdXMLclient_vars}
\title{The available variables for ndfdXMLclient client queries.}
\format{a list with 99 variable identifiers and default values
\describe{
}}
\source{
\url{http://graphical.weather.gov/xml/rest.php#XML_contents}
}
\usage{
ndfdXMLclient_vars
}
\description{
Variable are coded by short keywords. With the exceptions of \code{begin}, 
 \code{end} and \code{Unit} all are the values used in the examples at the
 source.  A complete list of \code{element} can be found here 
 \url{http://graphical.weather.gov/xml/docs/elementInputNames.php}
}
\keyword{datasets}

