% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{parameter_to_string}
\alias{parameter_to_string}
\title{Cast a parameter to character (suitable for DWMLDataRefClass$show)}
\usage{
parameter_to_string(x)
}
\arguments{
\item{x}{a list of one or more parameter lists}
}
\value{
character representation of the parameters
}
\description{
Cast a parameter to character (suitable for DWMLDataRefClass$show)
}

