% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{parse_response}
\alias{parse_response}
\title{Parse an \code{httr::response} class object to DWMLBaseRefClass object including
 an DWMLExceptionRefClass}
\usage{
parse_response(r, form = c("DWMLTopRefClass", "xml")[1], ...)
}
\arguments{
\item{r}{\code{httr::response} class object}

\item{form}{character specifies the type of data to return}

\item{...}{further arguments for \code{check_response()}}
}
\value{
object inheriting DWMLBaseRefClass
}
\description{
Parse an \code{httr::response} class object to DWMLBaseRefClass object including
 an DWMLExceptionRefClass
}

