% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{paste_listLatLon}
\alias{paste_listLatLon}
\title{Convert location pairs into characters suitable for 'listLatLon' 
 and 'gmlListLatLon'}
\usage{
paste_listLatLon(xy)
}
\arguments{
\item{xy}{matrix of [n,2] dims ([lat, lon] order)}
}
\value{
character in the form 'lat1,lon1 lat2,lon2 lat3,lon3 ...'
}
\description{
Convert location pairs into characters suitable for 'listLatLon' 
 and 'gmlListLatLon'
}

