% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_this}
\alias{query_this}
\title{Construct a query using an optional preconfigured list of defaults.}
\usage{
query_this(what = "point", element = c("mint", "maxt", "temp"), ...)
}
\arguments{
\item{what}{character - the type of query which defines the defaults.  Defaults
for each type of query are identified \code{ndfdXMLclient_vars} and 
\code{ndfdXMLclient_groups}
\itemize{
    \item point, Single Point Unsummarized Data
    \item multipoint, Multiple Point Unsummarized Data
    \item subgrid, Unsummarized Data for a Subgrid
    \item line, Unsummarized Data for a Line
    \item zipcodes, Unsummarized Data for One or More Zipcodes
    \item cities, Unsummarized Data for a List of Cities
    \item centerpoint, Unsummarized Data for a Subgrid Defined by a Center Point
    \item single_time, Unsummarized Data for a Single Time Encoded in dwGML
}}

\item{element}{a character vector or elements to retrieve.  See 
\url{http://graphical.weather.gov/xml/docs/elementInputNames.php}}

\item{...}{zero or more parameters that will add to or override the 
defaults listed in \code{ndfdXMLclient_vars}}
}
\value{
character of the query to pass to \code{get_query}
}
\description{
Defaults are stored in \code{ndfdXMLclient_vars} and are accessed by 
 the type of resources requested.  For instance the defaults for a 
 \code{point} are found in \code{ndfdXMLclient_groups} under \code{query_point}
}
\examples{
\dontrun{
# default for point, and then with lon and elements specified
query_this(what = "point")
query_this(what = "point", lon = -69, element = c("dew", "temp", "snow"))
# query by zipcode
query_this(what = "zipcode", zipCodeList = "04096",element = c("dew", "temp", "snow"))
}
}
\seealso{
\url{http://graphical.weather.gov/xml/rest.php#XML_contents}
}

