% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{split_indices}
\alias{split_indices}
\title{Groups indices (1 to n) into groups of MAX size}
\usage{
split_indices(index, MAX = 200)
}
\arguments{
\item{index}{numeric vector of indicies to split into vectors of MAX length}

\item{MAX}{numeric, the maximum number of elements per group returned}
}
\value{
list of index vectors of length MAX or shorter
}
\description{
Useful when needing to get data in groups of 200 as per \url{http://graphical.weather.gov/xml/rest.php}
}
\examples{
 \dontrun{
     str(split_indices(201:615))
 }
}

