% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{t2c}
\alias{t2c}
\title{Format date-time values into YYYY-mm-ddTHH:MM}
\usage{
t2c(x, fmt = "\%Y-\%m-\%dT\%H:\%M")
}
\arguments{
\item{x}{POSIXct of character time}

\item{fmt}{date format}
}
\value{
character time representation of input
}
\description{
Format date-time values into YYYY-mm-ddTHH:MM
}

