% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{which_client}
\alias{which_client}
\title{Undocumented query parameter (as far as I can tell) '?whichClient=xyz'}
\usage{
which_client(w = "query_point")
}
\arguments{
\item{w}{if not NA then retirve the client by this name, if NA return ALL}
}
\value{
a named character vector of client names or just one
}
\description{
Undocumented query parameter (as far as I can tell) '?whichClient=xyz'
}

