% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_value}
\alias{xml_value}
\title{Extract the value from a simple xml_node object}
\usage{
xml_value(x)
}
\arguments{
\item{x}{xml_node with a value}
}
\value{
the value of the node
}
\description{
Extract the value from a simple xml_node object
}

