% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xmlString}
\alias{xmlString}
\title{Convert XML::xmlNode to character}
\usage{
xmlString(x)
}
\arguments{
\item{x}{xmlNode or NodeRefClass}
}
\value{
character
}
\description{
Convert XML::xmlNode to character
}

