% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exception.R
\name{create_exception}
\alias{create_exception}
\title{Create an exception XML::xmlNode}
\usage{
create_exception(problem = "unspecified")
}
\arguments{
\item{problem}{character, the problem statement}
}
\value{
XML::xmlNode
}
\description{
Create an exception XML::xmlNode
}

