% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{get_one_parameter}
\alias{get_one_parameter}
\title{Extract one parameter from an xml_node}
\usage{
get_one_parameter(x)
}
\arguments{
\item{x}{xml_node}
}
\value{
a list with at least 'name' and 'value'
weather' = sapply(x[names(x) %in% 'weather-conditions'], xml2::xml_text),
conditions-icon' = sapply(x[names(x) %in% 'icon-link'], xml2::xml_text),
wordedForecast' = sapply(x[names(x) %in% 'text'], xml2::xml_txt),
}
\description{
Extract one parameter from an xml_node
}

