% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{get_one_parameter}
\alias{get_one_parameter}
\title{Extract one parameter from an XML::xmlNode}
\usage{
get_one_parameter(x)
}
\arguments{
\item{x}{XML::xmlNode}
}
\value{
a list with at least 'name' and 'value'
}
\description{
Extract one parameter from an XML::xmlNode
}

