% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_exception}
\alias{is_exception}
\title{Test XML::xmlNode or NodeRefClass is an exception}
\usage{
is_exception(x, space = "exc")
}
\arguments{
\item{space}{the namespace to test}

\item{node}{object to test}
}
\value{
logical
}
\description{
Test XML::xmlNode or NodeRefClass is an exception
}

