% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_atts}
\alias{xml_atts}
\title{Extract the attributes from a simple xml_node object}
\usage{
xml_atts(x)
}
\arguments{
\item{x}{xml_node with attributes}
}
\value{
character vector of the attributes
}
\description{
Extract the attributes from a simple xml_node object
}

