% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_name}
\alias{xml_name}
\title{Extract the name of a simple xml_node object}
\usage{
xml_name(x)
}
\arguments{
\item{x}{xml_node}

\item{character}{vector of the attributes}
}
\description{
Extract the name of a simple xml_node object
}

