% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{xml_value}
\alias{xml_value}
\title{Extract the value from a simple XML::xmlNode object}
\usage{
xml_value(x)
}
\arguments{
\item{x}{XML::xmlNode with a value}

\item{the}{value of the node}
}
\description{
Extract the value from a simple XML::xmlNode object
}

